+++ Data/SBV/Core/Symbolic.hs
@@ -2232,6 +2232,7 @@ data SMTConfig = SMTConfig {
 data TPOptions = TPOptions {
          ribbonLength :: Int  -- ^ Line length for TP proofs
        , quiet        :: Bool -- ^ No messages what-so-ever for successful steps. (Will print if something fails)
+       , printAsms    :: Bool -- ^ Print assumptions as they are proven as separate steps.
        , printStats   :: Bool -- ^ Print time/statistics. If quiet is True, then measureTime is ignored.
        , cacheProofs  :: Bool -- ^ Treat lemma names as unique, and cache the results. Default: False. Note that this
                               -- feature is unsound unless you make sure (by some other mechanism) that your lemma names